+++ AridArnold/Utility/MonoMath.cs
@@ -57,9 +57,9 @@
 		/// </summary>
 		public static Color Lerp(Color p1, Color p2, float t)
 		{
-			float R = (float)p1.R * (1.0f - t) + (float)p2.R * t;
-			float G = (float)p1.G * (1.0f - t) + (float)p2.G * t;
-			float B = (float)p1.B * (1.0f - t) + (float)p2.B * t;
+			float R = p1.R * (1.0f - t) + p2.R * t;
+			float G = p1.G * (1.0f - t) + p2.G * t;
+			float B = p1.B * (1.0f - t) + p2.B * t;
 
 			R = Math.Clamp(R, 0.0f, 255.0f);
 			G = Math.Clamp(G, 0.0f, 255.0f);
